UI/backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.30,<1.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
pydantic>=2.7,<3.0
minio>=7.2,<8.0
